Refactor file structure and update import paths: reorganize source files into 'Arbitrage Logic' directory, update import statements in index.js, test_cache_performance.js, and other relevant files. Add new puppeteer script to package.json for enhanced functionality. Log additional session exit timestamps in requests.log for better tracking.
WORK VERY WELL
