+++ requests.log
@@ -52,3 +52,5 @@
 {"type":"SESSION_EXIT","timestamp":"2025-08-13T18:33:16.300Z","info":{"lastExitAt":"2025-08-13T18:33:16.298Z"}}
 {"type":"SESSION_EXIT","timestamp":"2025-08-13T18:35:03.131Z","info":{"lastExitAt":"2025-08-13T18:35:03.129Z"}}
 {"type":"SESSION_EXIT","timestamp":"2025-08-13T18:36:39.780Z","info":{"lastExitAt":"2025-08-13T18:36:39.778Z"}}
+{"type":"SESSION_EXIT","timestamp":"2025-08-14T07:27:17.415Z","info":{"lastExitAt":"2025-08-14T07:27:17.414Z"}}
+{"type":"SESSION_EXIT","timestamp":"2025-08-14T07:41:47.093Z","info":{"lastExitAt":"2025-08-14T07:41:47.092Z"}}
