+++ src/Arbitrage Logic/utils/formatting.js
