+++ src/Arbitrage Logic/utils/performanceOptimizer.js
