+++ src/Arbitrage Logic/utils/validation.js
