+++ trades.log
@@ -1,0 +1,3 @@
+{"action":"ARBITRAGE_OPEN","symbol":"DEBT/USDT:USDT","timestamp":"2025-08-14T07:45:26.725Z","arbitrageId":"lbank-mexc-1755157526724-tfhxi","buyExchangeId":"lbank","sellExchangeId":"mexc","buyPrice":0.00215,"sellPrice":0.00223,"volume":5000,"buyAmount":5000,"sellAmount":5000,"buyCostUSD":10.75,"sellProceedsUSD":11.15,"diffPercent":"3.721%","totalInvestmentUSD":10.75,"expectedProfitUSD":0.40000000000000036,"tradingMode":"TOKEN","targetTokenQuantity":5000,"details":{"openTime":"2025-08-14T07:45:26.725Z","orderbookAtOpen":null,"profitBreakdown":{"grossDiffPercent":3.7209302325581493,"feesPercentTotal":0,"netExpectedDiffPercent":3.7209302325581493,"estimatedFeesUSD":0,"expectedGrossProfitUSD":0.40000000000000036,"expectedNetProfitUSD":0.40000000000000036,"buyNotionalUSD":10.75,"sellNotionalUSD":11.15,"totalNotionalUSD":21.9},"spreads":{"openDirection":"3.721%","oppositeDirection":null}}}
+{"action":"ARBITRAGE_OPEN","symbol":"DEBT/USDT:USDT","timestamp":"2025-08-14T07:45:33.387Z","arbitrageId":"lbank-mexc-1755157533387-lvcea","buyExchangeId":"lbank","sellExchangeId":"mexc","buyPrice":0.002152,"sellPrice":0.00223,"volume":5000,"buyAmount":5000,"sellAmount":5000,"buyCostUSD":10.76,"sellProceedsUSD":11.15,"diffPercent":"3.625%","totalInvestmentUSD":10.76,"expectedProfitUSD":0.39000000000000057,"tradingMode":"TOKEN","targetTokenQuantity":5000,"details":{"openTime":"2025-08-14T07:45:33.387Z","orderbookAtOpen":null,"profitBreakdown":{"grossDiffPercent":3.624535315985148,"feesPercentTotal":0,"netExpectedDiffPercent":3.624535315985148,"estimatedFeesUSD":0,"expectedGrossProfitUSD":0.39000000000000057,"expectedNetProfitUSD":0.39000000000000057,"buyNotionalUSD":10.76,"sellNotionalUSD":11.15,"totalNotionalUSD":21.91},"spreads":{"openDirection":"3.625%","oppositeDirection":null}}}
+{"action":"ARBITRAGE_OPEN","symbol":"DEBT/USDT:USDT","timestamp":"2025-08-14T07:45:49.985Z","arbitrageId":"lbank-mexc-1755157549985-t8ako","buyExchangeId":"lbank","sellExchangeId":"mexc","buyPrice":0.002149,"sellPrice":0.00223,"volume":5000,"buyAmount":5000,"sellAmount":5000,"buyCostUSD":10.745,"sellProceedsUSD":11.15,"diffPercent":"3.769%","totalInvestmentUSD":10.745,"expectedProfitUSD":0.40500000000000114,"tradingMode":"TOKEN","targetTokenQuantity":5000,"details":{"openTime":"2025-08-14T07:45:49.985Z","orderbookAtOpen":null,"profitBreakdown":{"grossDiffPercent":3.769194974406717,"feesPercentTotal":0,"netExpectedDiffPercent":3.769194974406717,"estimatedFeesUSD":0,"expectedGrossProfitUSD":0.40500000000000114,"expectedNetProfitUSD":0.40500000000000114,"buyNotionalUSD":10.745,"sellNotionalUSD":11.15,"totalNotionalUSD":21.895},"spreads":{"openDirection":"3.769%","oppositeDirection":null}}}
